requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
groq==0.26.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.13.2
Werkzeug==3.1.3
gunicorn==20.1.0
click==8.2.1
pymongo
PyPDF2
python-docx
flask_pymongo
flask_session
dash
plotly
fpdf
numpy
textblob
faiss-cpu
langchain-community
langchain_ollama
ollama
langdetect
Pandas
edge_tts
gtts